v3.2
- Criando o TaskContext, para cuidar especificamente das Tasks, para deixar o código mais fácil de entender. Aproveitando e excluindo conteúdo relacionado as tasks no IndexContext
